feat: add tosAcceptancePermitted flag to TapToPayConnectionConfiguration
This commit introduces a `tosAcceptancePermitted` flag to the `TapToPayConnectionConfiguration`. This flag controls whether the Terms of Service (ToS) acceptance prompt is shown during the connection process.

If set to `false`, the ToS prompt will not be displayed. This is useful for background connections. However, if the user has not previously accepted the ToS and this flag is `false`, the connection attempt will fail.

The default value for `tosAcceptancePermitted` is `true`.

This change affects the following files:
- `stripe_terminal/lib/src/platform/terminal_platform.api.dart`
- `stripe_terminal/ios/Classes/Mappings/ConnectionConfigurationMappings.swift`
- `stripe_terminal/ios/Classes/Api/TerminalApi.swift`
- `stripe_terminal/android/src/main/kotlin/mek/stripeterminal/api/TerminalApi.kt`
- `stripe_terminal/lib/src/models/connection_configuration.dart`